Optimize back end service logic
